+++ libuna/libuna_unicode_character.h
@@ -148,7 +148,7 @@ int libuna_unicode_character_copy_from_utf16_stream(
      const uint8_t *utf16_stream,
      size_t utf16_stream_size,
      size_t *utf16_stream_index,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -158,7 +158,7 @@ int libuna_unicode_character_copy_to_utf16_stream(
      uint8_t *utf16_stream,
      size_t utf16_stream_size,
      size_t *utf16_stream_index,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -193,7 +193,7 @@ int libuna_unicode_character_copy_from_utf32_stream(
      const uint8_t *utf32_stream,
      size_t utf32_stream_size,
      size_t *utf32_stream_index,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -203,7 +203,7 @@ int libuna_unicode_character_copy_to_utf32_stream(
      uint8_t *utf32_stream,
      size_t utf32_stream_size,
      size_t *utf32_stream_index,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 #if defined( __cplusplus )
