+++ libuna/libuna_utf32_stream.c
@@ -37,7 +37,7 @@ int libuna_utf32_stream_copy_byte_order_mark(
      uint8_t *utf32_stream,
      size_t utf32_stream_size,
      size_t *utf32_stream_index,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error )
 {
 	static char *function = "libuna_utf32_stream_copy_byte_order_mark";
@@ -217,7 +217,7 @@ int libuna_utf32_stream_size_from_utf8(
 int libuna_utf32_stream_copy_from_utf8(
      uint8_t *utf32_stream,
      size_t utf32_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      const libuna_utf8_character_t *utf8_string,
      size_t utf8_string_size,
      libcerror_error_t **error )
@@ -442,7 +442,7 @@ int libuna_utf32_stream_size_from_utf16(
 int libuna_utf32_stream_copy_from_utf16(
      uint8_t *utf32_stream,
      size_t utf32_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      const libuna_utf16_character_t *utf16_string,
      size_t utf16_string_size,
      libcerror_error_t **error )
@@ -629,7 +629,7 @@ int libuna_utf32_stream_size_from_utf32(
 int libuna_utf32_stream_copy_from_utf32(
      uint8_t *utf32_stream,
      size_t utf32_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      const libuna_utf32_character_t *utf32_string,
      size_t utf32_string_size,
      libcerror_error_t **error )
