+++ libuna/libuna_utf8_string.h
@@ -185,7 +185,7 @@ LIBUNA_EXTERN \
 int libuna_utf8_string_size_from_utf16_stream(
      const uint8_t *utf16_stream,
      size_t utf16_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      size_t *utf8_string_size,
      libcerror_error_t **error );
 
@@ -195,7 +195,7 @@ int libuna_utf8_string_copy_from_utf16_stream(
      size_t utf8_string_size,
      const uint8_t *utf16_stream,
      size_t utf16_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 /* The functionality for libuna_utf8_string_copy_to_utf16_stream is implemented by
@@ -209,7 +209,7 @@ int libuna_utf8_string_with_index_copy_from_utf16_stream(
      size_t *utf8_string_index,
      const uint8_t *utf16_stream,
      size_t utf16_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -218,7 +218,7 @@ int libuna_utf8_string_compare_with_utf16_stream(
      size_t utf8_string_size,
      const uint8_t *utf16_stream,
      size_t utf16_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -261,7 +261,7 @@ LIBUNA_EXTERN \
 int libuna_utf8_string_size_from_utf32_stream(
      const uint8_t *utf32_stream,
      size_t utf32_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      size_t *utf8_string_size,
      libcerror_error_t **error );
 
@@ -271,7 +271,7 @@ int libuna_utf8_string_copy_from_utf32_stream(
      size_t utf8_string_size,
      const uint8_t *utf32_stream,
      size_t utf32_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 /* The functionality for libuna_utf8_string_copy_to_utf32_stream is implemented by
@@ -285,7 +285,7 @@ int libuna_utf8_string_with_index_copy_from_utf32_stream(
      size_t *utf8_string_index,
      const uint8_t *utf32_stream,
      size_t utf32_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 LIBUNA_EXTERN \
@@ -294,7 +294,7 @@ int libuna_utf8_string_compare_with_utf32_stream(
      size_t utf8_string_size,
      const uint8_t *utf32_stream,
      size_t utf32_stream_size,
-     uint8_t byte_order,
+     int byte_order,
      libcerror_error_t **error );
 
 #if defined( __cplusplus )
